fix: resolve uint64_t cast ambiguity in Linux builds
Changed static_cast<uint64_t> to static_cast<unsigned long long> in
String::number() calls to resolve ambiguous overload resolution on
Linux platforms where uint64_t can match multiple integer types.

Fixes build errors in:
- STM32PinData.h:61 (makeSTM32Hash)
- PSOCPlatformConfiguration.cpp:677 (i2cSlaveTabName)

+++ source/library/include/qtac/STM32PinData.h
@@ -58,7 +58,7 @@ struct STM32PinData
 
     static HashType makeSTM32Hash(PinID gpioIndex)
     {
-        return strHash(qtac::String::number(static_cast<uint64_t>(gpioIndex)));
+        return strHash(qtac::String::number(static_cast<unsigned long long>(gpioIndex)));
     }
 
     HashType        _hash{0};
